refactor: split mod manager into focused services
